Refactor to share code for import modifier codemods
