+++ src/core_codemods/api/__init__.py
@@ -1,4 +1,4 @@
 # ruff: noqa: F401
 from codemodder.codemods.api import Metadata, Reference, ReviewGuidance
 
-from .core_codemod import CoreCodemod, SimpleCodemod
+from .core_codemod import CoreCodemod, ImportModifierCodemod, SimpleCodemod
